Refactoring: Separating DatabaseModelBuilder from DatabaseGenerator.
